Replace ChipTool with ChipServer and MatterYAMLRunner
